cmake: FindNagra: Allow using rel/dbg lib versions
Nagra is distributing libraries in release and debug variants. It may be
important during development to use the debug versions. There are
package config files for both of the variants, so make it possible to
use each in 'FindNagra.cmake'.

The decision about which variant to use is based on 'CMAKE_BUILD_TYPE',
debug variant for "Debug", release variant for anything else.

Signed-off-by: Ricardo Silva <ricardo.silva@gbtembedded.com>