cmake: FindNagra: Make Nagra include dirs optional
It's not mandatory to have a specific package include directory. The
standard sysroot include directory is fine, in which case the
'NAGRA_INCLUDE_DIRS' var will not be set. This is the case for the Nagra
package.

This way, remove 'NAGRA_INCLUDE_DIRS' from the required vars list in
'FIND_PACKAGE_HANDLE_STANDARD_ARGS', making it optional.

Signed-off-by: Ricardo Silva <ricardo.silva@gbtembedded.com>