platform: Use platform descrambling open/close functions
Descrambling sessions on the Nagra CMA library are managed via
nvDsmOpen() and nvDsmClose(). However, there are cases where
platform specific actions are required for the descrambling sessions to
be properly managed. This is the case of the Broadcom URSR (Unified
Reference Software Release) platform, for which this plugin is initially
being developed.

Unless some platform specific actions are done before nvDsmOpen(), it
won't then be possible to associate the key slots where the Nagra CMA
library writes the descrambling control words to the platform
infrastructure that deals with playback.

The 'nagra-cma' package's platform library is now exporting to new
functions:

 * nagra_cma_platf_dsm_open():
    Prepare a descrambling session on the platform.
    To be called before nvDsmOpen().

 * nagra_cma_platf_dsm_close():
    Terminate a descrambling session on the platform.
    To be called after nvDsmClose().

In the Broadcom URSR case, the implementation of this function relies on
the Broadcom's "CA Session Manager", designed specifically for this
goal.

This way, change MediaSessionSystem::OpenDescramblingSession() to call
nagra_cma_platf_dsm_open() before nvDsmOpen(). Conversely, also change
MediaSessionSystem::CloseDescramblingSession() to make the call to
nagra_cma_platf_dsm_close() after calling nvDsmClose().

On the platform functions, the descrambling session is identified by
TSID, so CloseDescramblingSession() needs it for the call to
nagra_cma_platf_dsm_close(). This way, change the prototype of
CloseDescramblingSession() to get by parameter the descrambling session
TSID maintained by the MediaSessionConnect object (the one that makes
the call to MediaSessionSystem::CloseDescramblingSession().

Signed-off-by: Ricardo Silva <ricardo.silva@gbtembedded.com>