MediaSystem: Store licenses persistently
The descrambling licenses are being successfully obtained and used.
However, their validity was only lasting as long as the application
session was open (between nvAsmOpen() and nvAsmClose()). Once the
application session got closed, the CMA requests a new license, even in
the case the license time window was still valid.

According to the Nagra CMA documentation, licenses can be made
persistent by using nvAsmUseStorage(), providing a directory where the
CMA can store the licenses persistently. This way, use this function for
that purpose.

To be noted that the license itself (even if stored persistently) has a
time window validity. The CMA will request for a new license once the
ones stored persistently are expired.

Signed-off-by: Ricardo Silva <ricardo.silva@gbtembedded.com>