MediaConnect: Add new PLATFORMDELIVERY update type
For achieving descrambling, it may be required to do some platform
specific actions besides feeding ECMs/EMMs via the
ECMDELIVERY/EMMDELIVERY update types. For example, it may be required to
provide information about player data, PIDs, etc., with the final goal
of making the correct association between the key slots where the CMA
writes the descrambling control words and the data that the player uses
for handling the AV streams.

This is the case of the Broadcom URSR (Unified Reference Software
Release) platform, for which this plugin is initially being developed,
where the player's playpump (or parserband) and respective PID channels
need to be associated with the descrambling session being managed by
this plugin.

The 'nagra-cma' package's platform library is now exporting a new
function named nagra_cma_platf_dsm_cmd() that allows providing different
types of data into a descrambling session identified by the TSID.

This way, add a new MediaSessionConnect update type, named
PLATFORMDELIVERY, to allow clients (e.g. player) provide the system data
that is required for the descrambling session to succeed.

The format of this data is up to the platform to specify (agreed between
the player side and the platform lib implementation side). This plugin
only has to pass it on transparently from a PLATFORMDELIVERY update
request into a call to nagra_cma_platf_dsm_cmd().

Since all the calls to Nagra CMA and platform library must be made from
the MediaSessionSystem (because there can only be one instance of the
Nagra static libs), extend it with a new SetPlatformMetadata() method
for delegating it the call to nagra_cma_platf_dsm_cmd() used for
servicing the new PLATFORMDELIVERY update request on
MediaSessionConnect.

Signed-off-by: Ricardo Silva <ricardo.silva@gbtembedded.com>